requirements.txt
fastapi
fastapi_jwt
fastapi_jwt_auth
passlib
PyJWT
pymysql
requests
uvicorn
snowflake-connector-python
pydantic[email]
pandas
gunicorn
snowflake-connector-python[pandas]
certifi==2020.12.5
click==7.1.2
fastapi==0.63.0
h11==0.11.0
httptools==0.1.1
mangum==0.10.0
pydantic==1.7.3
python-dotenv==0.15.0
python-json-logger==2.0.1
PyYAML==5.3.1
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.2
uvloop==0.14.0
watchgod==0.6
websockets==8.1
